Add a clean way to temporarily change the dev server proxy settings
vue.config.js will now check for the presence of a file `proxy-config.js`
in the `mocks` subdirectory.  If present, this file should export a proxy
configuration to use instead of the default, which remains
`http://localhost:5000`.  This file is gitignored.  Note that the dev
server must be restarted for the proxy config to change.